requirements.txt
Django
lxml
djangorestframework
markdown
django-filter
redis
rq
